requirements.txt
asgiref==3.5.2
autopep8==1.6.0
black==22.6.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.6
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
validate-email==1.3
whitenoise==6.2.0
